Survive a terminal narrower than the eighty-column layout
Shrinking the terminal below the header's node column - a phone with a
large font gets there at about forty columns - killed the reader: the
column positions are laid out for eighty columns, and width minus a
position past the right edge went to MakeAttrStr() as a size_t of four
billion. The positions are clamped to the window now, so a column past
the edge has no width rather than a negative one.

The message list had the same fault in another form: what the screen
has beyond eighty columns is shared out between the fields, and MAXCOL
being unsigned, a narrower screen wrapped the difference. Every field
was padded to a third of four billion spaces, vsnprintf() refused the
line and the program exited through ErrorExit.

The views are deleted before every window is closed on exit; done the
other way round, their close touched a window record already freed.
Found with AddressSanitizer, which is also what showed the two above
on a build that had survived them silently.

+++ golded3/gemlst.cpp
@@ -746,8 +746,14 @@ void GMsgList::Run()
 
     mlst_with_date(AA->Msglistdate());
 
-    fldadd1 = (MAXCOL-80)/3;
-    fldadd2 = (MAXCOL-80) - (fldadd1*2);
+    //  What the screen has beyond the eighty columns the layout is
+    //  drawn for, shared out between the fields - and taken from it
+    //  when the screen is narrower. MAXCOL is unsigned: on a terminal
+    //  under eighty columns the difference wrapped to four billion,
+    //  each field was padded to a third of that, and the line was
+    //  either killed by the memory it took or refused by vsnprintf().
+    fldadd1 = ((int)MAXCOL - 80) / 3;
+    fldadd2 = ((int)MAXCOL - 80) - (fldadd1 * 2);
 
     mlst = (MLst **)throw_malloc(sizeof(MLst *) * (maximum_index + 1));
